+++ rslang/.eslintrc.json
@@ -14,7 +14,7 @@
         "plugin:prettier/recommended",
         "prettier"
     ],
-    "parser": "@typescript-eslint/parser",
+    // "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaFeatures": {
             "jsx": true
