--- rslang/src/components/FormSigIn/FormSigIn.tsx
@@ -1,114 +1,0 @@
-import React, { useState } from "react";
-import "./FormSigIn.css";
-import Stack from "@mui/material/Stack";
-import Button from "@mui/material/Button";
-import {
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogContentText,
-  DialogActions,
-  TextField,
-} from "@mui/material";
-
-
-
-interface IFormSigin {
-  changeFormStateUser: () => void
-}
-
-const FormSigIn:React.FC<IFormSigin> = ({changeFormStateUser}) => {
-  const [open, setOpen] = React.useState(false);
-  const handleClose = () => {
-    setOpen(false);
-  };
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const [password, setPass] = useState("");
-  const [email, setEmail] = useState("");
-  const handleChangePassSigIN: React.ChangeEventHandler<
-    HTMLTextAreaElement | HTMLInputElement
-  > = (event) => {
-    setPass(event.target.value);
-  };
-  const handleChangeEmailSigIn: React.ChangeEventHandler<
-    HTMLTextAreaElement | HTMLInputElement
-  > = (event) => {
-    setEmail(event.target.value);
-  };
-
-
-  const [ localSigIn, setlocalSigIn ] = useState([]);
-
-  const handleSigIn = async () => {
-    setOpen(false);
-    const rawResponse = await fetch("https://rs-lang-1.herokuapp.com/signin", {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        email,
-        password,
-      }),
-    });
-    const content = await rawResponse.json();
-
-    let newArr: object[] = [...localSigIn, content]
-    localStorage.setItem("sigin", JSON.stringify(newArr));
-
-    console.log(content);
-    setPass("");
-    setEmail("");
-    changeFormStateUser()
-    console.log("Войти");
-  };
-
-  return (
-    <>
-      <Stack spacing={2} direction="row">
-        <Button variant="outlined" onClick={handleClickOpen}>
-          Войти
-        </Button>
-      </Stack>
-      <Dialog open={open} onClose={handleClose}>
-        <DialogTitle>Войти</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Для входа введите Ваш email и пароль
-          </DialogContentText>
-          <TextField
-            autoFocus
-            margin="dense"
-            id="email"
-            label="Введите Email"
-            type="email"
-            fullWidth
-            variant="standard"
-            value={email}
-            onChange={handleChangeEmailSigIn}
-          />
-          <TextField
-            margin="dense"
-            id="pass"
-            label="Введите пароль"
-            type="password"
-            fullWidth
-            variant="standard"
-            value={password}
-            onChange={handleChangePassSigIN}
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose}>Закрыть</Button>
-          <Button onClick={handleSigIn}>Войти</Button>
-        </DialogActions>
-      </Dialog>
-    </>
-  );
-}
-
-export default FormSigIn;
