--- rslang/src/components/index.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-
-import Form from './Form';
-
-ReactDOM.render(
-  <React.StrictMode>
-    <>
-      <Form/>
-    </>
-    
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-
-
