--- rslang/src/pages/Authorization.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Authorization() {
-  return <h1>Authorization Page</h1>;
-}
